message for delete

+++ src/components/list.js
@@ -5,7 +5,7 @@ import './list.css';
 
 const useStyles = makeStyles({
   TableContainer: {
-    width:700,
+    width: 700,
   },
   table: {
     width: 650,
@@ -27,50 +27,54 @@ function List(props) {
     }
     fetchEmployeeList();
   }, []);
-  const delEmp = async(id)=>{
-    console.log('eliminar empleados con id', id)
-    var requestOptions = {
-      method: 'DELETE',
-      redirect: 'follow'
-    };
+  const delEmp = async (id) => {
+    var isDelete = window.confirm("¿Eliminar este usuario?");
+    if (isDelete) {
+      console.log('eliminar empleados con id', id)
+      var requestOptions = {
+        method: 'DELETE',
+        redirect: 'follow'
+      };
 
-    await fetch(`https://my-test-cv.herokuapp.com/employee/${id}`, requestOptions)
-      .then(response => response.text())
-      .then(result => console.log(result))
-      .catch(error => console.log('error', error));
+      await fetch(`https://my-test-cv.herokuapp.com/employee/${id}`, requestOptions)
+        .then(response => response.text())
+        .then(result => console.log(result))
+        .catch(error => console.log('error', error));
+    } else {
+      console.log('ok')
+    }
   }
- 
+
   const classes = useStyles();
-  
 
   return (
     <TableContainer component={Paper}>
-    <Table className={classes.table}>
-      <thead>
-        <tr>
-          <th>Documento</th>
-          <th>Nombre</th>
-          <th>Apellido</th>
-          <th>Teléfono</th>
-          <th></th>
-          <th></th>
-        </tr>
-      </thead>
-       <tbody className="prop-list">
-         {employeeList.map(emp=>(
-           <tr key={emp.employee_id}>
-             <td>{emp.document_number}</td>
-             <td>{emp.first_name}</td>
-             <td>{emp.last_name}</td>
-             <td>{emp.phone_number}</td>
-             <td><Button variant="outlined" color="primary" onClick={()=>props.edit(emp.employee_id)}>Editar</Button></td>
-             <td><Button variant="outlined" color="primary" onClick={() => delEmp(emp.employee_id)}>x</Button></td>
-           </tr>
-         ))}
-       </tbody>
+      <Table className={classes.table}>
+        <thead>
+          <tr>
+            <th>Documento</th>
+            <th>Nombre</th>
+            <th>Apellido</th>
+            <th>Teléfono</th>
+            <th></th>
+            <th></th>
+          </tr>
+        </thead>
+        <tbody className="prop-list">
+          {employeeList.map(emp => (
+            <tr key={emp.employee_id}>
+              <td>{emp.document_number}</td>
+              <td>{emp.first_name}</td>
+              <td>{emp.last_name}</td>
+              <td>{emp.phone_number}</td>
+              <td><Button variant="outlined" color="primary" onClick={() => props.edit(emp.employee_id)}>Editar</Button></td>
+              <td><Button variant="outlined" color="primary" onClick={() => delEmp(emp.employee_id)}>x</Button></td>
+            </tr>
+          ))}
+        </tbody>
       </Table>
     </TableContainer>
-    );
+  );
 }
